MADE: --> Revised TestGroundEKF.cpp and TestSegmentation.cpp looking for the error. Not found. --> Going to test with simulation (To "investigate" if error is in BOViL class or in the data set conversion

+++ Tests/test_groundEKF/source/TestGroundEKF.cpp
@@ -142,8 +142,8 @@ void testSegmentation(){
 
 		cv::imshow("ORIGINAL", ori);
 
-		double arrayZk[2] = {	float (objects[maxIndex].getCentroid().x),
-								float (objects[maxIndex].getCentroid().y)};
+		double arrayZk[2] = {	double (objects[maxIndex].getCentroid().x),
+								double (objects[maxIndex].getCentroid().y)};
 
 		groundEKF.stepEKF(BOViL::math::Matrix<double>(arrayZk, 2, 1), inputBuffer[0] - lastTime);
 		lastTime = inputBuffer[0];
